Удаление ненужных более файлов проекта, связанных с реализацией архитектуры MVP. Интерфейс ViewModel.kt сделан отдельно, вынесен из LoginContract.kt, который впоследствии удалён. Устранение замечаний по части codeStyle - переименование некоторых элементов макета для выполнения соглашения по именованию.
