--- app/src/main/java/ru/profitsw2000/mvpapp/ui/Constants.kt
@@ -1,7 +1,0 @@
-package ru.profitsw2000.mvpapp.ui
-/*
-
-const val ERROR_SIGN_IN = 1
-const val ERROR_PASSWORD_RESTORE = 2
-const val ERROR_SIGN_UP = 3
-const val ERROR_EMPTY_FIELD = 4*/
