+++ app/src/main/res/layout/activity_account.xml
@@ -8,7 +8,7 @@
     tools:context=".ui.screens.AccountActivity">
 
     <TextView
-        android:id="@+id/tv_login_succesful"
+        android:id="@+id/login_successful_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/login_successful_text"
