+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -52,7 +52,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/tv_restore_password_succesful"
+        android:id="@+id/restore_password_successful_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/dialog_restore_password_success_text"
