+++ app/src/main/res/layout/activity_sign_up.xml
@@ -77,7 +77,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/tv_sign_up_succesful"
+        android:id="@+id/sign_up_successful_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/tv_sign_up_success"
